Move `Config` to `build` package
