+++ tools/src/main/scala/scala/scalanative/optimizer/Reporter.scala
@@ -9,16 +9,18 @@ import scalanative.io.{withScratchBuffer, VirtualDirectory}
 trait Reporter {
 
   /** Gets called whenever optimizations starts. */
-  private[scalanative] def onStart(batchId: Int, batchDefns: Seq[Defn]): Unit = ()
+  private[scalanative] def onStart(batchId: Int, batchDefns: Seq[Defn]): Unit =
+    ()
 
   /** Gets called right after pass transforms the batchDefns. */
   private[scalanative] def onPass(batchId: Int,
-             passId: Int,
-             pass: Pass,
-             batchDefns: Seq[nir.Defn]): Unit = ()
+                                  passId: Int,
+                                  pass: Pass,
+                                  batchDefns: Seq[nir.Defn]): Unit = ()
 
   /** Gets called with final result of optimization. */
-  private[scalanative] def onComplete(batchId: Int, batchDefns: Seq[Defn]): Unit = ()
+  private[scalanative] def onComplete(batchId: Int,
+                                      batchDefns: Seq[Defn]): Unit = ()
 }
 
 object Reporter {
