+++ tools/src/main/scala/scala/scalanative/optimizer/inject/ClassStruct.scala
@@ -20,6 +20,6 @@ class ClassStruct(top: Top) extends Inject {
 }
 
 object ClassStruct extends InjectCompanion {
-  override def apply(config: tools.Config, top: Top) =
+  override def apply(config: build.Config, top: Top) =
     new ClassStruct(top)
 }
