+++ tools/src/main/scala/scala/scalanative/optimizer/pass/AsLowering.scala
@@ -23,6 +23,6 @@ class AsLowering extends Pass {
 }
 
 object AsLowering extends PassCompanion {
-  override def apply(config: tools.Config, top: Top) =
+  override def apply(config: build.Config, top: Top) =
     new AsLowering
 }
