+++ tools/src/main/scala/scala/scalanative/optimizer/pass/Canonicalization.scala
@@ -51,6 +51,6 @@ class Canonicalization extends Pass {
 }
 
 object Canonicalization extends PassCompanion {
-  override def apply(config: tools.Config, top: Top) =
+  override def apply(config: build.Config, top: Top) =
     new Canonicalization
 }
