+++ tools/src/main/scala/scala/scalanative/optimizer/pass/ClosureLowering.scala
@@ -19,6 +19,6 @@ class ClosureLowering extends Pass {
 }
 
 object ClosureLowering extends PassCompanion {
-  override def apply(config: tools.Config, top: Top) =
+  override def apply(config: build.Config, top: Top) =
     new ClosureLowering
 }
