+++ tools/src/main/scala/scala/scalanative/optimizer/pass/ConstantFolding.scala
@@ -141,7 +141,7 @@ class ConstantFolding extends Pass {
 }
 
 object ConstantFolding extends PassCompanion {
-  override def apply(config: tools.Config, top: Top) =
+  override def apply(config: build.Config, top: Top) =
     new ConstantFolding
 
   object IVal {
