+++ tools/src/main/scala/scala/scalanative/optimizer/pass/CopyPropagation.scala
@@ -67,6 +67,6 @@ class CopyPropagation extends Pass {
 }
 
 object CopyPropagation extends PassCompanion {
-  override def apply(config: tools.Config, top: Top) =
+  override def apply(config: build.Config, top: Top) =
     new CopyPropagation()
 }
