+++ tools/src/main/scala/scala/scalanative/optimizer/pass/MethodLowering.scala
@@ -61,6 +61,6 @@ class MethodLowering(implicit top: Top) extends Pass {
 }
 
 object MethodLowering extends PassCompanion {
-  override def apply(config: tools.Config, top: Top) =
+  override def apply(config: build.Config, top: Top) =
     new MethodLowering()(top)
 }
