+++ tools/src/main/scala/scala/scalanative/optimizer/pass/PartialEvaluation.scala
@@ -231,7 +231,7 @@ class PartialEvaluation extends Pass {
 }
 
 object PartialEvaluation extends PassCompanion {
-  override def apply(config: tools.Config, top: Top) =
+  override def apply(config: build.Config, top: Top) =
     new PartialEvaluation
 
   object PowerOf2 {
