+++ tools/src/main/scala/scala/scalanative/optimizer/pass/SizeofLowering.scala
@@ -26,6 +26,6 @@ class SizeofLowering(top: Top) extends Pass {
 }
 
 object SizeofLowering extends PassCompanion {
-  override def apply(config: tools.Config, top: Top) =
+  override def apply(config: build.Config, top: Top) =
     new SizeofLowering(top)
 }
