+++ tools/src/main/scala/scala/scalanative/optimizer/pass/TypeValueLowering.scala
@@ -19,6 +19,6 @@ class TypeValueLowering(implicit top: Top) extends Pass {
 }
 
 object TypeValueLowering extends PassCompanion {
-  override def apply(config: tools.Config, top: Top) =
+  override def apply(config: build.Config, top: Top) =
     new TypeValueLowering()(top)
 }
