+++ tools/src/main/scala/scala/scalanative/optimizer/pass/UnitSimplification.scala
@@ -18,6 +18,6 @@ class UnitSimplification extends Pass {
 }
 
 object UnitSimplification extends PassCompanion {
-  override def apply(config: tools.Config, top: Top) =
+  override def apply(config: build.Config, top: Top) =
     new UnitSimplification
 }
