+++ tools/src/test/scala/scala/scalanative/CodeGenSpec.scala
@@ -3,7 +3,7 @@ package scala.scalanative
 import java.nio.file.{Path, Paths}
 import scalanative.io.VirtualDirectory
 import scalanative.optimizer.Driver
-import scalanative.tools.Config
+import scalanative.build.Config
 import scalanative.util.Scope
 
 /** Base class to test code generation */
