+++ tools/src/test/scala/scala/scalanative/LinkerSpec.scala
@@ -5,9 +5,10 @@ import scala.language.implicitConversions
 import java.io.File
 import java.nio.file.{Files, Path, Paths}
 
+import build.Config
 import util.Scope
 import nir.Global
-import tools.{Config, Mode}
+import tools.Mode
 import optimizer.Driver
 
 import org.scalatest.FlatSpec
