+++ tools/src/test/scala/scala/scalanative/optimizer/DriverSpec.scala
@@ -1,7 +1,7 @@
 package scala.scalanative
 package optimizer
 
-import tools.Config
+import build.Config
 import analysis.ClassHierarchy.Top
 
 import org.scalatest._
