+++ tools/src/test/scala/scala/scalanative/optimizer/pass/AsLoweringTest.scala
@@ -3,6 +3,7 @@ package optimizer
 package pass
 
 import analysis.ClassHierarchy.Top
+import build.Config
 import nir._
 import tools._
